# internal/core/server/openapi.yaml
openapi: 3.0.3
info:
  title: Feather Feature Store API
  description: |
    Feather is a high-performance, real-time feature store providing sub-millisecond
    feature retrieval through a tiered storage architecture.

    ## Authentication
    API requests can be authenticated using API keys passed in the `X-API-Key` header.

    ## Rate Limiting
    Requests are rate-limited per client IP using a token bucket algorithm.
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

    ## Request Tracing
    Include `X-Request-ID` header for distributed tracing. If not provided, one will be generated.
  version: 1.0.0
  contact:
    name: Feather Support
    url: https://github.com/feather-store/feather
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://feather.example.com
    description: Production
...
paths:
...
  /v1/features:
    get:
      tags: [Features]
      summary: Get features for an entity
      description: Retrieve feature values for a specific entity
      operationId: getFeatures
      parameters:
        - name: entity
          in: query
          required: true
          description: Entity identifier (e.g., "user:12345")
          schema:
            type: string
          example: "user:12345"
        - name: features
          in: query
          required: false
          description: Comma-separated list of feature names. If omitted, returns all features.
          schema:
            type: string
          example: "click_count,last_active,purchase_total"
      responses:
        '200':
          description: Features retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFeaturesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Features]
      summary: Store features for an entity
      description: Store or update feature values for an entity
      operationId: putFeatures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutFeaturesRequest'
      responses:
        '200':
          description: Features stored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/features/batch:
    post:
      tags: [Features]
      summary: Batch get features for multiple entities
      description: Retrieve features for multiple entities in a single request
      operationId: batchGetFeatures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGetRequest'
      responses:
        '200':
          description: Features retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/features/history:
    get:
      tags: [Features]
      summary: Get historical features (point-in-time)
      description: Retrieve feature values as they existed at a specific point in time
      operationId: getFeaturesAsOf
      parameters:
        - name: entity
          in: query
          required: true
          description: Entity identifier
          schema:
            type: string
          example: "user:12345"
        - name: features
          in: query
          required: false
          description: Comma-separated list of feature names
          schema:
            type: string
        - name: as_of
          in: query
          required: true
          description: Point-in-time timestamp (RFC3339 format)
          schema:
            type: string
            format: date-time
          example: "2024-01-15T10:30:00Z"
      responses:
        '200':
          description: Historical features retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFeaturesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # Schema Operations
  # ============================================
  /v1/schema/groups:
    get:
      tags: [Schema]
      summary: List all feature groups
      description: Retrieve all registered feature group schemas
      operationId: listFeatureGroups
      responses:
        '200':
          description: Feature groups retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFeatureGroupsResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Schema]
      summary: Create a feature group
      description: Register a new feature group schema
      operationId: createFeatureGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureGroup'
      responses:
        '201':
          description: Feature group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Feature group already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/schema/groups/{name}:
    get:
      tags: [Schema]
      summary: Get a feature group
      description: Retrieve a specific feature group schema by name
      operationId: getFeatureGroup
      parameters:
        - name: name
          in: path
          required: true
          description: Feature group name
          schema:
            type: string
          example: "user_features"
      responses:
        '200':
          description: Feature group retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureGroupResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Schema]
      summary: Delete a feature group
      description: Remove a feature group schema (does not delete stored data)
      operationId: deleteFeatureGroup
      parameters:
        - name: name
          in: path
          required: true
          description: Feature group name
          schema:
            type: string
      responses:
        '200':
          description: Feature group deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # Vector Operations
  # ============================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        request_id:
          type: string
          example: "req_abc123"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "Entity ID is required"
        request_id:
          type: string
          example: "req_abc123"
    FeatureValue:
      type: object
      properties:
        value:
          oneOf:
            - type: integer
            - type: number
            - type: string
            - type: boolean
            - type: array
              items:
                type: number
          description: The feature value
        timestamp:
          type: string
          format: date-time
          description: When the value was recorded
        version:
          type: integer
          format: int64
          description: Version number for optimistic locking

    # ============================================
    # Feature Schemas
    # ============================================
    GetFeaturesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            entity_id:
              type: string
              example: "user:12345"
            features:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/FeatureValue'
        request_id:
          type: string
    PutFeaturesRequest:
      type: object
      required:
        - entity_id
        - features
      properties:
        entity_id:
          type: string
          example: "user:12345"
        features:
          type: object
          additionalProperties:
            oneOf:
              - type: integer
              - type: number
              - type: string
              - type: boolean
          example:
            click_count: 42
            last_active: "2024-01-15T10:30:00Z"
    BatchGetRequest:
      type: object
      required:
        - entities
      properties:
        entities:
          type: array
          items:
            type: string
          maxItems: 100
          example: ["user:12345", "user:67890"]
        features:
          type: array
          items:
            type: string
          example: ["click_count", "last_active"]
    BatchGetResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            results:
              type: object
              additionalProperties:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/FeatureValue'
        request_id:
          type: string

    # ============================================
    # Schema Definitions
    # ============================================
    FeatureGroup:
      type: object
      required:
        - name
        - entity_type
        - features
      properties:
        name:
          type: string
          example: "user_features"
        entity_type:
          type: string
          example: "user"
        description:
          type: string
        ttl:
          type: string
          description: Duration string (e.g., "24h", "7d")
          example: "24h"
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureSpec'
        tags:
          type: object
          additionalProperties:
            type: string
    FeatureSpec:
      type: object
      required:
        - name
        - data_type
      properties:
        name:
          type: string
          example: "click_count"
        data_type:
          type: string
          enum: [int64, float64, string, bool, bytes, vector, timestamp]
        description:
          type: string
        aggregation:
          $ref: '#/components/schemas/AggregationSpec'
        validation:
          $ref: '#/components/schemas/ValidationSpec'
    AggregationSpec:
      type: object
      properties:
        function:
          type: string
          enum: [count, sum, avg, min, max, last]
        window:
          type: string
          example: "1h"
        slide_by:
          type: string
          example: "5m"
    ValidationSpec:
      type: object
      properties:
        min:
          type: number
        max:
          type: number
        not_null:
          type: boolean
        one_of:
          type: array
          items:
            type: string
        regex:
          type: string
    ListFeatureGroupsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            groups:
              type: array
              items:
                $ref: '#/components/schemas/FeatureGroup'
        request_id:
          type: string
    FeatureGroupResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/FeatureGroup'
        request_id:
          type: string

    # ============================================
    # Vector Schemas
    # ============================================
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - ApiKeyAuth: []